travel-iter/pyproject.toml
[project]
name = "travel-iter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fpdf>=1.7.2",
    "langchain-community>=0.3.26",
    "langgraph>=0.4.8",
    "python-dotenv>=1.1.0",
    "streamlit>=1.46.0",
]
